pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "alpaca-options-bot"
version = "0.1.0"
description = "Automated options trading bot with Alpaca integration, backtesting, and terminal UI"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "Developer" }]
keywords = ["trading", "options", "alpaca", "backtesting", "finance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "alpaca-py>=0.21.0",
    "rich>=13.7.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "aiohttp>=3.9.0",
    "python-dateutil>=2.8.0",
    "pytz>=2024.1",
    "pyyaml>=6.0.0",
    "scipy>=1.11.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "typer>=0.9.0",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
    "black>=24.1.0",
    "pre-commit>=3.6.0",
]
backtest = [
    "pyarrow>=14.0.0",
    "plotly>=5.18.0",
]
all = [
    "alpaca-options-bot[dev,backtest]",
]

[project.scripts]
alpaca-options = "alpaca_options.cli:main"

[project.urls]
Homepage = "https://github.com/username/alpaca-options-bot"
Documentation = "https://github.com/username/alpaca-options-bot#readme"
Repository = "https://github.com/username/alpaca-options-bot"

[tool.hatch.build.targets.wheel]
packages = ["src/alpaca_options"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function calls in argument defaults
    "B905",   # zip strict
]

[tool.ruff.lint.isort]
known-first-party = ["alpaca_options"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "alpaca.*",
    "scipy.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra -q --cov=alpaca_options --cov-report=term-missing"

[tool.coverage.run]
source = ["src/alpaca_options"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

# UV configuration
[tool.uv]
managed = true
package = true

# Dependency groups (replaces dev-dependencies)
[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
    "pre-commit>=3.6.0",
    "pyarrow>=14.0.0",
    "plotly>=5.18.0",
]
